Fix 32/64 bit break in ControlNetwork binary format
I incorrectly used size_t for a property in ControlMeasure. That type
changes size depending on platform.

+++ src/vw/BundleAdjustment/ControlNetwork.cc
@@ -50,7 +50,7 @@ namespace ba {
   /// Constructor
   ControlMeasure::ControlMeasure( float col, float row,
                                   float col_sigma, float row_sigma,
-                                  size_t image_id,
+                                  uint64 image_id,
                                   ControlMeasureType type ) :
     m_col(col), m_row(row), m_col_sigma(col_sigma), m_row_sigma(row_sigma), m_image_id(image_id), m_type(type) {
